Persistent contact information
Contact information now persists between 'Add Contact' activities, which
will allow it to be committed to memory. This makes full use of the
InputHolder class, which is now implemented as a Singleton pattern
class, allowing the instance to be called statically from any of the
activities. InputHolder now has a destroy method that will be used to
de-instantiate it at the end of the AddContact sequence.

Changed int numbers to Strings in Contact/InputHolder. These will be
converted to ints at a different stage.