Contact Detail editable
The first value (First name) in Contact detail is now editable and will
update immediately in both the Contact detail screen upon closing the
dialog box, as well as being updated in the ListView when returning.

Also, the functionality of the back button in the Contact detail page
has been added, and whenever the ListView is resumed, it will be
refreshed.